Sprint 10 / Homework / Fix 1 : extracted dimensions from layouts, moved classes to files, moved mock to object
