+++ app/src/main/java/com/practicum/playlistmaker/TrackAdapter.kt
@@ -1,0 +1,14 @@
+package com.practicum.playlistmaker
+
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+
+class TrackAdapter(val trackList: List<Track>) : RecyclerView.Adapter<TrackViewHolder>() {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int) =
+        TrackViewHolder(parent)
+    override fun getItemCount(): Int = trackList.size
+    override fun onBindViewHolder(holder: TrackViewHolder, position: Int) {
+        holder.bind(trackList[position])
+    }
+
+}
