+++ app/src/main/res/values/styles.xml
@@ -32,6 +32,10 @@
                 <item name="android:fontFamily">@font/ys_text_medium</item>
                 <item name="android:textFontWeight">500</item>
                 <item name="android:textSize">22sp</item>
+                <item name="android:layout_width">match_parent</item>
+                <item name="android:layout_height">0dp</item>
+                <item name="android:layout_marginBottom">16dp</item>
+                <item name="android:layout_weight">1</item>
             </style>
 
         <style name="Widget.PlayListMaker.CommonActivity">
